requirements.txt
psutil==5.9.0
supabase>=1.0.0
python-dotenv

fastapi
uvicorn[standard]
flask
flask-socketio
flask-cors
python-socketio
